test(node): Migrate logging integration tests to create*Tests helpers
Move the winston, pino and consola node integration test suites onto the
shared runner helpers, splitting each `subject.ts`/inline server into an
`instrument.mjs` (Sentry.init) + `scenario.mjs` pair.

- pino: `createEsmTests` (keeps its existing ESM-only coverage)
- winston, consola: `createCjsTests` (keeps their existing CJS-only coverage)

No coverage change vs. the previous per-suite run mode; each suite runs in
the same module format it did before.
