+++ dev-packages/node-integration-tests/suites/consola/scenario-args.mjs
@@ -1,16 +1,7 @@
 import * as Sentry from '@sentry/node';
-import { loggingTransport } from '@sentry-internal/node-integration-tests';
 import { consola } from 'consola';
 
-Sentry.init({
-  dsn: 'https://public@dsn.ingest.sentry.io/1337',
-  release: '1.0.0',
-  environment: 'test',
-  enableLogs: true,
-  transport: loggingTransport,
-});
-
-async function run(): Promise<void> {
+async function run() {
   // Set consola level to capture all logs including debug and trace
   consola.level = 5;
